Fixed the issue with the BloomFilter. The hash functions did not round the hash value to a integer value correctly.

+++ src/main/java/pt/it/av/atnog/utils/structures/bloom/Hash.java
@@ -5,5 +5,5 @@ package pt.it.av.atnog.utils.structures.bloom;
  * @version 1.0
  */
 public interface Hash<T> {
-  long hash(T e);
+  int hash(T e);
 }
